Update streamlit_app.py
made changes to put the list of selected fruits into a variable called fruits_selected. Then, we ask our app to use the fruits in our fruits_selected list to pull rows from the full data set (and assign that data to a variable called fruits_to_show). Finally, we ask the app to use the data in fruits_to_show in the dataframe it displays on the page.